requirements.txt
flask
mediapipe
opencv-python
numpy
scipy
scikit-learn
matplotlib
plotly
pandas
fpdf2
